Finished Products page - Added eslint for Javascript code, properly setup for VueJS + Prettier - Modified deploy workflow - Run Deploy steps only on branch 'master' - Return custom Error page with 404 when route is not found

+++ webpack.mix.js
@@ -38,6 +38,10 @@ mix
         require('postcss-nested'),
         require('tailwindcss')
     ])
-    .purgeCss()
-    .version()
     .sourceMaps()
+
+if (mix.inProduction()) {
+    mix
+        .purgeCss()
+        .version()
+}
